Add .pre stage to GitLab CI
It was discovered that in some condition the cache is not working as it
should and the jobs are failing as it can not find some files. To be
more precise, each job required to do npm install and then it was
possible to run the second command.

Right now, the install stage with npm install is going to be run once
and then all the next stages will use it.

Reference:
https://stackoverflow.com/questions/55359804/gitlab-ci-npm-doesnt-like-the-cached-node-modules